chore(playwright): extract shared chat mock helpers
Move `buildMockStream`, `buildMockImageGenStream`, `buildMockSearchStream`,
and the chat-endpoint route helpers into `utils/chatMock.ts` so specs no
longer duplicate them. Unify `mockChatEndpoint` on a pre-built-body
signature; call sites in `chat_message_rendering.spec.ts` now wrap
explicitly with `buildMockStream`.

Also fix a strict-mode violation in the code preview modal test where
`getByText("greet")` matched both the `greet.py` filename and the
highlighted `greet` function token.
